backend/requirements.txt
# Core Flask Dependencies
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
Werkzeug==3.1.3

# Authentication
PyJWT==2.10.1

# Audio Processing
librosa==0.11.0
soundfile==0.13.1
pyloudnorm==0.1.1
audioread==3.0.1

# AI/ML for Audio Analysis
openai-whisper==20231117
torch==2.5.1
torchaudio==2.5.1
google-genai==1.0.0
shazamio==0.6.0

# PDF Generation
reportlab==4.4.4
fpdf==1.7.2

# HTTP Client
requests==2.32.5
httpx==0.28.1

# Utilities
python-dotenv==1.0.0
